Refactor: remove ActivitiesResponse from CardsRestClient and add the error param to the existing ActivitiesResponse. Update unit tests
